hmm. this is awkward.
gf2x includes its impl and config headers in the "small" multiplication
files it exports. This inconditionally exposes and redefines the
PACKAGE__* and such variables for the user. This is bad.

cantor.h uses (was using) the gf2xfft-config.h ; as a fix to the previous
one, we have a relatively easy time getting rid of it, since it's equally
bad.

cantor.h includes mpfq/<whichever>.h though, and that includes inline
assembly code. If gf2x is available, it makes sense to use the gf2x
optimized routine in this case. But HAVE_GF2X_H_ is defined in
gf2xfft-config.h only.

and mpfq/XXXX.h pretending a one-size-fits-all implementation suffices is
really problematic for fownstream users.