still work in progress -- but postponed to later
make check doesn't correctly recurse in the fft/ subdir.
I'm not terribly happy with the names so far.
the tests in fft/Makefile.am would need to be done about the same way
they're done in tests/Makefile.am, with externally-generated checksum.
We'd need to make the generation code a common base.